tools/scylla-nodetool: do not check if rpc_server is up
thrift support was removed in ad649be1, the API for checking /
starting / stopping rpc_server (thrift_server) was removed as well.
so, the existing clients using this API would have 404 when querying
it.

so in this change, let's just hardwire the result to "false" instead
of querying this removed RESTful API.

the test is updated accordingly. since in new installations, we
still package the java-based nodetool, which still queries this API,
this API is preserved for the backward compatiblity.

Fixes scylladb/scylladb#19923
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>