tablets: locator: Allow specifying tablet count in when allocating for new table
The use case is to allocate a similar tablet map to the existing one
using the same allocation logic which is used for a new table.

This will be used by load balancer in the mode which performs
migrations which drive existing map to the perfect one achieved
by allocating the table from scratch with the same tablet count.