topology_coordinator: introduce reload_count in topology state and use it to prevent race
Topology request table may change between the code reading it and
calling to cv::when() since reading is a preemption point. In this
case cv:signal can be missed. Detect that there was no signal in between
reading and waiting by introducing reload_count which is increased each
time the state is reloaded and signaled. If the counter is different
before and after reading the state may have change so re-check it again
instead of sleeping.

Fixes: scylladb/scylladb#19994
(cherry picked from commit 3ca8bdea11814f7a2218701a2d203f10c68b0d95)