table: Add CRU and CWU counters to the table stats
This patch adds the ability to track consumption units

It adds the The `consumption_unit_counter` class which encapsulates the
concept of consumption units, which are simplified representations of user
operations within the system.

Each operation consumes a specific number of units based on its
complexity and resource requirements.

This class is front-end protocol-agnostic; it does not depend on any specific
protocol. Each protocol that uses this class will define
how the units are updated and managed within its context.

The patch adds two counters, cru (Consumption Read Units) and cwu
(Consumption Write Units), to measure read and write operations in a
standardized way using consumption_unit_counter.

Currently, these counters will be utilized by Alternator, with plans to
extend their use to CQL in the future.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>