test_return_consumedcapacity.py: Add putItem tests
This patch adds testing for putItem consume capacity.

The basic test now checks both get and put consume capacity return
values.

There is an additional test for number support. Numbers are encoded
differently with alternator and dynamoDB, the test adds some flexibility
in the result.
So it would pass both.

`