utils/div_ceil: sanity check to abort when integer overflow
integer overflow can occur when adding or subtracting two integers,
particularly when the result exceeeds the maximum or the minimum
value that can be represented by the integer data type of the result.
so, in this change, we check and abort the application if we run into
such integer overflow.

Refs #20000
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>