build: do not regenerate building system for seastar or abseil
before this change, we regenerate the building systems for seastar and
abseil even if their building systems, i.e., `build.ninja` have been
generated. this works fine alebit that if, for instance, seastar has
been built, after regenerating `build.ninja`, all source files are
rebuilt. this is a waste of time. because to update the build
dependencies of a certain .o file, ninja needs to regenerate the
corresponding .o.d file, but we just generate the .o and .o.d in
a single pass, so the source files are always recompiled even when
it's not necessary.

in this change, if the `build.ninja` file exists in the build directory,
instead of rerunning `cmake` to generate this file, we just skip the
step to generate the building system. as the `build.ninja` script
generated by CMake is able to regenerate itself when its dependencies
are updated, and the rule for generating it is named `RERUN_CMAKE`,
which just reruns `cmake` with cached command line arguments passed to
this tool in previous run. this helps us to avoid the unnecessary
compilation of seastar and abseil after regenerating `build.ninja`
using `configure.py`.

Fixes scylladb/scylladb#20079
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>