test: perf: tablet_load_balancing: Create proper replication strategy for tables in the schema
The new load balancer algorithm will consult the table's replication
strategy, so we need the correct replication strategy in the databaes
schema, which uses intended replication factor. To create tables using
such replication strategy, we need enough nodes in topology, so the
test needs to now create the correct topology in group0.