service/migration_manager: do not move away a const reference
despite that `std::move()` makes its argument a rvalue reference,
in this case, its argument is already a lvalue reference. what
`std::move(view)` returns is a
`std::remove_reference<const schema_ptr&>::type`, i.e,
`const schema_ptr`.

so we cannot benefit from "moving away" from it, and what gets called
is the copy constructor of `schema_ptr` when calling
`view_ptr(std::move(view))`.

so, in this change, we make this more obvious by

* changing the type of `view` to `const auto&`
* remove the `std::move()`. it's but confusing.

this change is a cleanup to improve the readability.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>