doc: add info about file-based streaming to a separate file
This commit adds a file with information about file-based streaming to the _common folder.
The file is then included using the scylladb_include_flag directive.

The major purpose of this commit is to make it possible to include a different file
in the scylla-enterprise repo - with different content (feature documentation).

Refs https://github.com/scylladb/scylla-enterprise/issues/4585
Refs https://github.com/scylladb/scylla-enterprise/issues/4254