sstable: use full scan reader in make_reader() when appropriate
there are two cases when loading a given sstable:

- load the full range and the whole paritition
- load a partial of the range and/or partial of the partition

in the first case, we do not need to seek to the given token, so
the full scan reader is a better fit. and per testing, it has slightly
better performance than the regular ranged reader in the "loading"
phase when performing load-n-stream operation

| reader    | mean (partitions / sec) | stderr (partitions / sec) |
| --------- | ----------------------- | ------------------------- |
| range     | 514692.33               | 3502.87                   |
| full_scan | 558914.36               | 4547.98                   |

but we also perform full-range load when performing compaction. so
compaction can also benefit from using the full scan reader.

in this change, full scan reader is used when performing a full-range
scan for better performance of full-range load-n-stream and compaction.
and otherwise falls back to the ranged reader

Refs scylladb/scylladb#19989

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>