storage_service: get ERM pointer to pause load balancing
In some operations (like truncate), it is necessary to pause load
balancing until the operation is complete to avoid resurrected data.
This patch adds a function to storage_service which waits for any
running load balancing for a given table to complete, then returns an
Effective Replication Map pointer. Any subsequent load balancing will be
paused until the pointer is released.
In case the table for which we are waiting does not use tablets,
the wait is immediatelly satisfied and the function will return.